add test cases of lab 3 - problem 2

+++ simultaneous-grader/src/main/java/model/resultReport/ProblemResultDetails.java
@@ -2,7 +2,7 @@ package model.resultReport;
 
 import model.component.Student;
 
-public record ProblemResultDetails(String name, Student student, double passedPercent) {
+public record ProblemResultDetails(String name, Student student, int passedPercent) {
 	public void toNotFoundRecord() {
 		
 	}
